Did some code refactoring and removed unwanted dependancies
