+++ pom.xml
@@ -9,10 +9,10 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
-	<artifactId>GitHubActions</artifactId>
+	<artifactId>GitHubActionsQWas</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>GitHubActions</name>
-	<description>Demo project for Spring Boot GithubActions</description>
+	<name>GitHubActionsQWas</name>
+	<description>GitHub-Action-QWas</description>
 	<properties>
 		<java.version>17</java.version>
 	</properties>
@@ -49,11 +49,6 @@
 			<artifactId>junit</artifactId>
 			<version>4.13.2</version>
 			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.google.guava</groupId>
-			<artifactId>guava</artifactId>
-			<version>32.1.3-jre</version>
 		</dependency>
 	</dependencies>
