+++ src/main/java/com/example/GitHubActionsQWas/WASClient/InvalidAPIResponseException.java
@@ -1,4 +1,4 @@
-package com.example.GitHubActions.WASClient;
+package com.example.GitHubActionsQWas.WASClient;
 
 public class InvalidAPIResponseException extends Exception {
 }
