+++ src/main/java/com/example/GitHubActionsQWas/util/Helper.java
@@ -1,4 +1,4 @@
-package com.example.GitHubActions.util;
+package com.example.GitHubActionsQWas.util;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
