--- src/test/java/com/example/GitHubActions/GitHubActionsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.GitHubActions;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GitHubActionsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
